epoll `wait` performance improvement and relaxed `HTTPRouterProtocol` requirements
