+++ Sources/DestinyDefaults/router/DestinyHTTPMutableRouterProtocol.swift
@@ -1,0 +1,5 @@
+
+import DestinyBlueprint
+
+public protocol DestinyHTTPMutableRouterProtocol: DestinyHTTPRouterProtocol, HTTPMutableRouterProtocol {
+}
